workflow/testing: add network mode runtime test (#123)
* task run: Support network boot
* set default provisioning url to qemu host ip
* set default network mode to dhcp
* start python http.server if boot mode is "network"
* only add local ospkgs if boot mode is set to local
* workflow/testing: Add network mode runtime test

+++ .github/workflows/scripts/test-qemu.sh
@@ -2,25 +2,21 @@
 
 set -Eeuo pipefail
 
-[ "$#" -ne 1 ] && echo "run: ${0} <mbr/efi>" && exit 1
-
 TIMEOUT=300
 LOG=
-TARGET=
+TARGET=${1:-mbr}
 IMAGE=
-case $1 in
+case $TARGET in
 	"mbr")
 		LOG=/tmp/mbr-qemu.log
-		TARGET=run-mbr
 		IMAGE=stboot_mbr_installation.img
 		;;
 	"efi")
 		LOG=/tmp/efi-qemu.log
-		TARGET=run-efi
 		IMAGE=stboot_efi_installation.img
 		;;
 	*)
-		echo "unknown installation: $1."
+		echo "unknown installation: $TARGET."
 		exit 1
 		;;
 esac
@@ -35,14 +31,16 @@ MATCH+=("Debian GNU/Linux 10 debian ttyS0")
 
 cleanup () {
 	qemu_pid=$(pgrep -f "qemu-system-x86_64.*${IMAGE}")
-	kill -TERM "${qemu_pid}"
+	[ -z "$qemu_pid" ] || kill -TERM "${qemu_pid}"
 	pkill -TERM -P $$
 }
 
-trap cleanup EXIT
+trap cleanup 0
 
 # run qemu
-task ${TARGET} </dev/null | tee /dev/stderr > "$LOG" &
+source st.config
+ST_LOCAL_OSPKG_DIR=$ST_LOCAL_OSPKG_DIR ST_BOOT_MODE=$ST_BOOT_MODE ./scripts/qemu_run.sh \
+    -b "${TARGET}" -i "out/${IMAGE}" </dev/null | tee /dev/stderr > "$LOG" &
 
 i=0
 while [ "$i" -lt "$TIMEOUT" ]
